Added Health Bar + Script
Will change to on collision with damaging/healing items, but keybinds are Space (damage) and Q (health)-
Steam system will be similar if we plan to keep it in the game. There is also a pause script wip but its yet to be finalized out, so far I hope this is alright!